Move portable biomes system to own class
